requirements.txt
atriumdb[mariadb]==2.1.1
aiomysql==0.2.0
aio_pika>=8.3.0
apscheduler==3.9.1.post1
databases==0.7.0
fastapi==0.101.1
fastapi_auth0==0.3.2
numpy>=1.21
opentelemetry-distro==0.36b0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-instrumentation-fastapi==0.40b0
orjson>=3.6
pydantic==1.10.7
PyYAML==6.0
requests==2.31.0
setuptools~=57.0.0
siridb-connector==2.1.3
SQLAlchemy==1.4.46
uvicorn==0.23.2
xxhash~=3.2.0